ref: move LocationType to own file
